Move utilities into inspec. Update the conversion from inspec to xccdf for newer inspec runs.
